Bug fixes
Fixed Byte being a 64 bit unsigned integer :-/
Multiple changes to the SpriteKit engine to support texture caching
Updated message publication in SwiftScript engine to use the correct method of generating the event
Added an engine instance to DecodingContext

+++ Sources/Pelota/Data Model/BinaryRepresentations.swift
@@ -6,4 +6,4 @@
 
 import Foundation
 
-public typealias Byte = UInt
+public typealias Byte = UInt8
